+++ module/sheets/GSActorSheet.mjs
@@ -434,7 +434,7 @@ export default class GSActorSheet extends ActorSheet{
 				if(x === 1){
 					for(const [key, item] of Object.entries(results[1])){
 						if(key === 'recovery' || key === 'power')
-							diceHold = this._setSpellPowerDice(key, item, spellUsed, tempTargets);
+							diceHold = this._setSpellPowerDice(key, item, spellUsed, tempTargets, rollTotal);
 					}
 				}
 				if(x === 0)
@@ -937,17 +937,17 @@ export default class GSActorSheet extends ActorSheet{
 		<div class="target grid grid-9col">
 			<img class="targetImg" src="${activeTarget.document.texture.src}">
 			<h3 class="targetName grid-span-5">${activeTarget.document.name}</h3>
-			<button type="button" class="monsterDefRoll gm-section" data-monsterid="${monster._id}" data-playerid="${this.actor._id}" data-type="dodge" data-value="${dodgeValue}" title="${game.i18n.localize('gs.dialog.actorSheet.itemsTab.dodge')}"><i class="fa-solid fa-angles-right"></i></button>
-			<button type="button" class="monsterDefRoll gm-section" data-monsterid="${monster._id}" data-playerid="${this.actor._id}" data-type="block" data-value="${blockValue}" title="${game.i18n.localize('gs.dialog.actorSheet.itemsTab.block')}" ${hasBlock?``:`disabled`}><i class="fa-solid fa-shield-halved"></i></i></button>
+			<button type="button" class="monsterDefRoll gm-view" data-monsterid="${monster._id}" data-playerid="${this.actor._id}" data-type="dodge" data-value="${dodgeValue}" title="${game.i18n.localize('gs.dialog.actorSheet.itemsTab.dodge')}"><i class="fa-solid fa-angles-right"></i></button>
+			<button type="button" class="monsterDefRoll gm-view" data-monsterid="${monster._id}" data-playerid="${this.actor._id}" data-type="block" data-value="${blockValue}" title="${game.i18n.localize('gs.dialog.actorSheet.itemsTab.block')}" ${hasBlock?``:`disabled`}><i class="fa-solid fa-shield-halved"></i></i></button>
 			<button type="button" class="actorDamageRoll" data-extradmg="${extraDmg}" data-playerid="${this.actor._id}" data-id="${itemInfo._id}" title="${game.i18n.localize('gs.dialog.actorSheet.itemsTab.power')}"><i class="fa-solid fa-burst"></i></button>
 		</div>`;
 		return targetMessage;
 	}
 
-	_setSpellPowerDice(key, extractedDice, spell, targets){
+	_setSpellPowerDice(key, extractedDice, spell, targets, spellDC){
 		return `<div class='spellTarget grid grid-2col'>
 			<div style="display:flex; justify-content: center; align-items: center; font-size: 14px;">${game.i18n.localize('gs.dialog.spells.rolldice')}</div>
-			<button type="button" class="actorSpellDmg" data-targets="${targets}" data-keytype="${key}" data-rolldice="${extractedDice}" data-playerid="${this.actor._id}" data-spell="${spell._id}"><i class="fa-solid fa-wand-magic-sparkles"></i></button>
+			<button type="button" class="actorSpellDmg" data-spelldc="${spellDC}" data-targets="${targets}" data-keytype="${key}" data-rolldice="${extractedDice}" data-playerid="${this.actor._id}" data-spell="${spell._id}"><i class="fa-solid fa-wand-magic-sparkles"></i></button>
 		</div>`;
 	}